# backend/docs/swagger.yaml
definitions:
  backend_internal_entities.Subject:
    properties:
      id:
        type: string
      nameKz:
        type: string
      nameRu:
        type: string
      slug:
        type: string
    type: object
  internal_adapters_http_handlers.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  internal_adapters_http_handlers.CreateCourseRequest:
    properties:
      description:
        type: string
      difficulty_level:
        maximum: 5
        minimum: 1
        type: integer
      subject_id:
        type: string
      title:
        type: string
    required:
    - difficulty_level
    - subject_id
    - title
    type: object
  internal_adapters_http_handlers.CreateCourseResponse:
    properties:
      id:
        type: string
    type: object
  internal_adapters_http_handlers.CreateLessonRequest:
    properties:
      content_text:
        type: string
      module_id:
        type: string
      order_index:
        type: integer
      title:
        type: string
      video_url:
        type: string
    required:
    - module_id
    - order_index
    - title
    type: object
  internal_adapters_http_handlers.CreateLessonResponse:
    properties:
      lesson_id:
        type: string
    required:
    - lesson_id
    type: object
  internal_adapters_http_handlers.CreateModuleRequest:
    properties:
      course_id:
        type: string
      order_index:
        type: integer
      title:
        type: string
    required:
    - course_id
    - order_index
    - title
    type: object
  internal_adapters_http_handlers.CreateModuleResponse:
    properties:
      module_id:
        type: string
    required:
    - module_id
    type: object
  internal_adapters_http_handlers.ErrorResponse:
    properties:
      message:
        example: something went wrong
        type: string
    type: object
  internal_adapters_http_handlers.GetStructureResponse:
    properties:
      modules:
        items:
          $ref: '#/definitions/internal_adapters_http_handlers.ModuleResponse'
        type: array
    type: object
  internal_adapters_http_handlers.LessonResponse:
    properties:
      content_text:
        type: string
      file_attachment_url:
        type: string
      id:
        type: string
      module_id:
        type: string
      order_index:
        type: integer
      title:
        type: string
      video_url:
        type: string
      xp_reward:
        type: integer
    type: object
  internal_adapters_http_handlers.LoginRequest:
    properties:
      email:
        example: student@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  internal_adapters_http_handlers.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_adapters_http_handlers.ModuleResponse:
    properties:
      course_id:
        type: string
      id:
        type: string
      lessons:
        items:
          $ref: '#/definitions/internal_adapters_http_handlers.LessonResponse'
        type: array
      order_index:
        type: integer
      title:
        type: string
    type: object
  internal_adapters_http_handlers.PublishStatusRequest:
    properties:
      is_published:
        type: boolean
    type: object
  internal_adapters_http_handlers.RegisterResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
  internal_adapters_http_handlers.ResetPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - email
    - new_password
    - old_password
    type: object
  internal_adapters_http_handlers.UpdateCourseRequest:
    properties:
      cover_image_url:
        type: string
      description:
        type: string
      difficulty_level:
        type: integer
      title:
        type: string
    type: object
  internal_adapters_http_handlers.UpdateModuleRequest:
    properties:
      order_index:
        type: integer
      title:
        type: string
    type: object
  internal_adapters_http_handlers.UploadResponse:
    properties:
      url:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for the School Backend service.
  title: School Backend API
  version: "1.0"
paths:
  /v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change password for the currently authenticated user
      parameters:
      - description: Password change payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - multipart/form-data
      description: Register a new user (student or teacher)
      parameters:
      - description: User Avatar Image
        in: formData
        name: avatar
        type: file
      - description: User Email
        in: formData
        name: email
        required: true
        type: string
      - description: User Password (min 8 chars)
        in: formData
        name: password
        required: true
        type: string
      - description: User Role (student or teacher)
        in: formData
        name: role
        required: true
        type: string
      - description: First Name
        in: formData
        name: first_name
        required: true
        type: string
      - description: Last Name
        in: formData
        name: last_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Change password using email and old password
      parameters:
      - description: Reset password payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
      summary: Reset password (public)
      tags:
      - auth
  /v1/courses:
    post:
      consumes:
      - application/json
      description: Create a course (Teacher/Admin only)
      parameters:
      - description: Course data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.CreateCourseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a new course
      tags:
      - courses
  /v1/courses/{id}:
    put:
      consumes:
      - application/json
      description: Update title, description, cover, etc. (Author only)
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update course details
      tags:
      - courses
  /v1/courses/{id}/publish:
    post:
      consumes:
      - application/json
      description: Change visibility of the course
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Status
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.PublishStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Publish or Unpublish course
      tags:
      - courses
  /v1/courses/{id}/structure:
    get:
      description: Get modules and lessons for editor
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.GetStructureResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get full course structure
      tags:
      - courses
  /v1/lessons:
    post:
      consumes:
      - application/json
      parameters:
      - description: Lesson data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.CreateLessonRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.CreateLessonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Add lesson to module
      tags:
      - lessons
  /v1/modules:
    post:
      consumes:
      - application/json
      parameters:
      - description: Module data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.CreateModuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.CreateModuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Add module to course
      tags:
      - modules
  /v1/modules/{id}:
    delete:
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete module
      tags:
      - modules
    put:
      consumes:
      - application/json
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      - description: Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_handlers.UpdateModuleRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update module
      tags:
      - modules
  /v1/subjects:
    get:
      description: Get list of available subjects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_entities.Subject'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
      summary: Get all subjects
      tags:
      - subjects
  /v1/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload file to MinIO (avatar, course_cover, lesson_material)
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: File type (avatar, cover, lesson)
        in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.UploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a file
      tags:
      - upload
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"